+++ HarZindagi/app/src/main/java/com/ipal/itu/harzindagi/Activities/RegisteredChildActivity.java
@@ -143,13 +143,13 @@ public class RegisteredChildActivity extends BaseActivity {
 
                 AlertDialog.Builder adb = new AlertDialog.Builder(RegisteredChildActivity.this);
 
-                adb.setTitle("Want To Delete User");
+                adb.setTitle("کیاآپ بچے کاڈیٹا ختم کرناچاہتے ہیں؟");
 
 
                 adb.setIcon(R.drawable.info_circle);
 
 
-                adb.setPositiveButton("Yes", new DialogInterface.OnClickListener() {
+                adb.setPositiveButton("ہاں", new DialogInterface.OnClickListener() {
                     public void onClick(DialogInterface dialog, int which) {
                         final List<ChildInfo> data = ChildInfoDao.getByKIdAndIMEI(childID, imei);
                         List<KidVaccinations> items=KidVaccinationDao.getById(data.get(0).kid_id);
@@ -161,7 +161,11 @@ public class RegisteredChildActivity extends BaseActivity {
 
                         if (data.size()>0)
                         {
-                            bookses.get(0).delete();
+                            if(bookses.size()>0)
+                            {
+                                bookses.get(0).delete();
+                            }
+
                             data.get(0).delete();
                             Intent in=new Intent(RegisteredChildActivity.this,DashboardActivity.class);
                             startActivity(in);
@@ -174,7 +178,7 @@ public class RegisteredChildActivity extends BaseActivity {
                 });
 
 
-                adb.setNegativeButton("No", new DialogInterface.OnClickListener() {
+                adb.setNegativeButton("نہیں", new DialogInterface.OnClickListener() {
                     public void onClick(DialogInterface dialog, int which) {
 
                         dialog.dismiss();
